requirements.txt
argparse==1.4.0
catboost==1.2
category_encoders==2.6.0
einops==0.6.0
kaleido==0.2.1
libzero==0.0.3.dev7
lightgbm==3.3.5
Markdown==3.4.1
matplotlib==3.7.1
opencv-contrib-python==4.7.0.72
openml==0.13.0
openpyxl==3.1.2
optuna==3.1.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.4.0
plotly==4.14.3
protobuf==3.19.6
pydantic==1.10.6
Pygments==2.14.0
pytomlpp==1.0.11
pytorch-lightning==1.8.6
pytorch-tabnet==4.0
PyYAML==5.4.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
shap==0.41.0
slicer==0.0.7
spacy==3.5.0
tensorboard==2.12.0
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.3
torchvision==0.14.1
tqdm==4.65.0
wandb==0.13.11
wasabi==1.1.1
wget==3.2
xgboost==1.7.5
